BE-Flask/requirements.txt
boto == 2.49.0
boto3 == 1.26.66
moviepy == 1.0.3
Flask == 2.2.2
Flask-Cors == 3.0.10
python-dotenv == 0.21.1